Model Of This Project
This is the model of this project . I will be using Tkinter for implementing the Graphical User Interface for this question. The basic agenda of this project is to create an interactive python game in which the user will have to guess the word based on a series of questions.